removed smi package and moved smi to adapter package
