+++ adapter/error.go
@@ -15,6 +15,8 @@
 package adapter
 
 import (
+	"fmt"
+
 	"github.com/layer5io/meshkit/errors"
 )
 
@@ -71,4 +73,24 @@ func ErrNewSmi(err error) error {
 
 func ErrRunSmi(err error) error {
 	return errors.NewDefault(ErrRunSmiCode, "Error running SMI conformance test", err.Error())
+}
+
+// ErrSmiInit is the error for smi init method
+func ErrSmiInit(des string) error {
+	return errors.NewDefault(errors.ErrSmiInit, des)
+}
+
+// ErrInstallSmi is the error for installing smi tool
+func ErrInstallSmi(err error) error {
+	return errors.NewDefault(errors.ErrInstallSmi, fmt.Sprintf("Error installing smi tool: %s", err.Error()))
+}
+
+// ErrConnectSmi is the error for connecting to smi tool
+func ErrConnectSmi(err error) error {
+	return errors.NewDefault(errors.ErrConnectSmi, fmt.Sprintf("Error connecting to smi tool: %s", err.Error()))
+}
+
+// ErrDeleteSmi is the error for deleting smi tool
+func ErrDeleteSmi(err error) error {
+	return errors.NewDefault(errors.ErrDeleteSmi, fmt.Sprintf("Error deleting smi tool: %s", err.Error()))
 }
